requirements.txt
django
pillow
djangorestframework
python-binance
APScheduler